face-service/requirements.txt
# Face Recognition Microservice Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3

# Face Recognition & Computer Vision
face_recognition==1.3.0
opencv-python==4.9.0.80
numpy==1.26.3
Pillow==10.2.0

# Anti-Spoofing with DeepFace
deepface==0.0.89
tensorflow>=2.13.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0

# For development
pytest==7.4.4
httpx==0.26.0
